+++ ReleaseNotes.md
@@ -1,4 +1,21 @@
 # Release Notes
+### v0.3.2.0
+- **New Features**
+  - **Export Comparison Report**
+    - When comparing against a baseline, you can now export the report as CSV or Markdown.
+  - **New Application Types for Assignments**
+    - androidForWorkApp
+    - androidLobApp
+    - androidStoreApp
+    - iosLobApp
+    - iosVppApp
+
+- **Bug Fixes & Improvements**
+  - **Assignment Creation**
+    - Fixed an issue when no security groups are present in the tenant.
+  - **Refactoring Code**
+    - Moved funtions to function File
+
 ### v0.3.1.0
 - **New Features**
   - **Confirmation Pop-Ups**